// ModernizacionPersonas.Web/ClientApp/src/app/models/informacion-intermediario.ts
export interface InformacionIntermediario {
  codigoCotizacion: number;
  codigo: number;
  clave?: number;
  participacion: number;
  primerNombre: string;
  segundoNombre: string;
  primerApellido: string;
  segundoApellido: string;
  codigoTipoDocumento: number;
  numeroDocumento: string;
  email: string;
  codigoEstado: number;
}